src/requirements.txt
# Framework
Django==2.1.1

# Python server
gunicorn==19.7.1

# PostgreSQL library
psycopg2==2.7.3.2

# Social network authorization
social-auth-app-django==2.1.0

# Image library
Pillow==4.2.1

# Deletes from /media when object deleted (not done by default)
django-cleanup==1.0.1

# Markdown in templates
django-markdown-deux==1.0.5

# Models can be made sortable in Admin Panel
django-admin-sortable2==0.6.21

# Model version are stored separately, so they can be restored
django-reversion==3.0.0

# BBCode
django-precise-bbcode==1.2.8

# xlsx & csv terminal viewer
x-x==0.9

# xlsx reader
openpyxl==2.4.8

# Vk API (get user id by short link)
vk==2.0.2

# Show progress bar in scripts
progressbar2==3.34.3

# Timezone support
pytz==2017.2

# Simple schedules
django-background-tasks==1.1.12

# File cloud
django-filebrowser-no-grappelli==3.7.4
